updated models in database and created store for portfolio and API

+++ python-project-starter/app/models/portfolio.py
@@ -8,7 +8,7 @@ class Portfolio(db.Model):
     user_id = db.Column(db.Integer, db.ForeignKey("users.id"), nullable=False)
     ticker = db.Column(db.String, nullable=False)
     quantity = db.Column(db.Integer)
-    average_price = db.Column(db.Numeric(10, 2))
+    average_price = db.Column(db.Float)
 
     user = db.relationship("User", back_populates="portfolio")
